Move eventhistory code directly into profile.jsx
